ai_server/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
pytesseract==0.3.10
Pillow==10.4.0
numpy==1.26.4
opencv-python==4.9.0.80
google-generativeai==0.5.3
gunicorn==21.2.0
